Implemented some other DD API calls
