+++ lib/hotdog/commands/untag.rb
@@ -1,7 +1,5 @@
 #!/usr/bin/env ruby
 
-require "fileutils"
-
 module Hotdog
   module Commands
     class Untag < BaseCommand
@@ -48,46 +46,21 @@ module Hotdog
           if options[:tags].empty?
             # delete all user tags
             with_retry do
-              detach_tags(host, source=options[:tag_source])
+              @source_provider.detach_tags(host, source=options[:tag_source])
             end
           else
-            host_tags = with_retry { host_tags(host, source=options[:tag_source]) }
+            host_tags = with_retry { @source_provider.host_tags(host, source=options[:tag_source]) }
             old_tags = host_tags["tags"]
             new_tags = old_tags - options[:tags]
             if old_tags == new_tags
               # nop
             else
               with_retry do
-                update_tags(host, new_tags, source=options[:tag_source])
+                @source_provider.update_tags(host, new_tags, source=options[:tag_source])
               end
             end
           end
         end
-      end
-
-      private
-      def detach_tags(host_name, options={})
-        code, detach_tags = dog.detach_tags(host_name, options)
-        if code.to_i / 100 != 2
-          raise("dog.detach_tags(#{host_name.inspect}, #{options.inspect}) returns [#{code.inspect}, #{detach_tags.inspect}]")
-        end
-        detach_tags
-      end
-
-      def host_tags(host_name, options={})
-        code, host_tags = dog.host_tags(host_name, options)
-        if code.to_i / 100 != 2
-          raise("dog.host_tags(#{host_name.inspect}, #{options.inspect}) returns [#{code.inspect}, #{host_tags.inspect}]")
-        end
-        host_tags
-      end
-
-      def update_tags(host_name, tags, options={})
-        code, update_tags = dog.update_tags(host_name, tags, options)
-        if code.to_i / 100 != 2
-          raise("dog.update_tags(#{host_name.inspect}, #{tags.inspect}, #{options.inspect}) returns [#{code.inspect}, #{update_tags.inspect}]")
-        end
-        update_tags
       end
     end
   end
